+++ resources/views/components/page-header.blade.php
@@ -5,8 +5,8 @@
 ])
 
 <div {{ $attributes->merge(['class' => 'mb-8']) }}>
-    <h1 class="text-5xl font-medium text-white">{{ $title }}</h1>
+    <h1 class="text-5xl font-medium ui-text">{{ $title }}</h1>
     @if($description)
-        <p class="text-sm text-white/40 mt-2">{{ $description }}</p>
+        <p class="text-sm ui-text-subtle mt-2">{{ $description }}</p>
     @endif
 </div>